✨ Feature: Reintroduce Syntax class for custom syntax highlighting
